authentication configuration is completed.

+++ src/main/java/com/enkuru/springtry/repository/HashTagRepository.java
@@ -11,5 +11,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  * Time: 15:25
  */
 @RepositoryRestResource
-public interface HashTagRepository extends JpaRepository<HashTag, Integer> {
+public interface HashTagRepository extends JpaRepository<HashTag, Long> {
 }
